requirements.txt
pydantic==1.10.9
pandas==1.3.4
openpyxl==3.1.2
jinja2==3.0.3
python-multipart==0.0.6
aiofiles==0.8.0
firebase-admin==5.0.2
alembic==1.11.1
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.98.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0